Share arcam_fmj convert_exception decorator from entity module
Hoists the ConnectionFailed→HomeAssistantError decorator out of the
media_player module so the upcoming remote platform can apply the same
translation. Behavior is unchanged.
